+++ tests/Zend/Stdlib/AllTests.php
@@ -36,7 +36,7 @@ require_once 'Zend/Stdlib/SplPriorityQueueTest.php';
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  * @group      Zend_Stdlib
  */
-class Zend_EventManager_AllTests
+class Zend_Stdlib_AllTests
 {
     public static function main()
     {